Dance Sequencer(WIP)
Working on a script that can record and store dance movements, so that it can be played back.
It records the Vector3 positions of the object being controlled, 16 beats
(the position on each beat)
The plan is to use this sequence and string it together with tweens or lerps, in order to "play back" the recorded animation.