Script for VOX switching
Wow this was some brain gymnastics
There are different voice cues unique to each character
But every game, the lines of dialogue can be spoken by completely different characters
So had to code a way for the game to know which character is speaking, and which line should be played.
Sounds simple. It wasn't.