Geometry Shader Explodes Triangles
Managed to make the geometry shader explode the triangle.
It took a while, but I found the correct shader library to use so that I could do the world space to clip space conversion.

For cursed geometry, it could be good to mix up the different spaces, like object, , world, clip...
Also messing with views calculations