Repeller Class
Took the previous "effect point" out into its own class, renamed it as "repeller" (because it describes the visual effect)
At the moment im just using an object's position in world space to create this "repeller", but I could later find other logic to set a repeller point more dynamically through code. This could be caused by things in real time, like proximity of other objects, or ray calculations from the view/camera, etc.
This depends on what is desired ito visual disruption or "cursebending": how does it mess with player's expectations of the gameworld's reality/ontology? How does it affect the perception/experience/cybernetic relation to the character? How does it draw attention to the procedural process at work in our interactive maniplulation with the animation?
I don't want to merely create "visual effects" - but intentionally draw attention to the procedural processes that construct the visual reality that we interact with and expand the boundaries of the "playground" outside the diegetic reality of the narrative to include the actual procesesses into the "digital dance".