Grid Spacing Feature, Opaque Texture
Added a feature that will space all vertices to snap to a kind of grid.
The size of the spacing can be adjusted, so the grid has tinier detail or more blocky detail.
I'm quite pleased with this outcome, I think it "draws attention" to the contructedness of the mesh, which is kind of the aim of all these features.

I want to make sure I dont just make cool VFX, like showoff shaders. It has to serve a purpose. And somehow my aim is to "queer" the experience of interactive animation. Drawing attention to artifice/contruntion is only a part of it, I think it can be pushed into something more interesting though...