Colors, Irridescence
Spent most of the day wokring Colors into my script. Note to future self: Set up Colors at the start when the script is generated!!! As well as other things like smoothing and metallic etc!!! Because adding it into every need packing and struct in EVERY PASS is super super painful!

ANyway, manged to get color in everywhere it was needed, so the started playing with it.

I am super inspired to do an irridescent shader. Managed to get the basic idea started with a simple "hard Fresnel".

Next I would like to work out how to do a falloff and a color gradient, to use in the two areas seperated by the hard fresnel divide.