Setup for ShaderBending
I have made a basic movement control script to control the character.
I am aiming to make a shader that has some kind of "trails/tails" behaviour.
My plan is to use the control object postitions to affect a bit of mesh displacement.
There is actually 3 objects for each control object, following with damp transform.
So, using 3 transform positions, I could do some kind of bezier curve math...

What Im trying to work towards in general, is to have shader effects respond to controller input, so that the glitchyness can be part of a player-character cybernetic loop. I want the glitch and jank to actually affect the "performance" of this cybernetic process. So things like shader effects should serve this disruption, not just be a "cool effect"