Starting Over with Shader
Shader was crashing Unity, so I'm starting over.
This time I'm setting up color and other properties in the shader graph BEFORE is start work on geometery shader.

There was some issue with the shader overloading the gpu
Of course I dont understand why.
Something about infinite loops etc.

The take-away is I need to build it with a bit more care to not create loops in places i don't understand, and be careful about using EXPENSIVE processes