Cubify
Now I used the "cubify" effect that gets created by the control objects (the controlled bones)
The effect points have an area of effect and a falloff, which sets the amount of "cubification".
(It's basically just snapping vertices to incremented x,y,z values, so it aligns to a kind of 3d matrix... relative to the falloff amount)

I'm pretty chuffed with this effect at the moment.

The next step is gonna be to use the movement speed to add to the effect, perhaps affecting the geometry in another way.
I somehow want a trailing feeling that makes the interaction visually interesting to play with.. but also CURSED