Geometry and MoveDirection
some script changes to make the geometry effetcs respond in relation to movement direction.
I tried for a while to use rigidbody velocity, but need more time to learn how it works.
I am using follow objects and checking the direction between them.
So then i can use that direction as a factor to change the falloff (e.g. stronger effect in the direction of movement)
and also the direction of things like displacement