CursedShader1
I think I'm happy with the current Cursed shader.
It can have a range of glitchyness, and I think it should be one of maybe a few variations of "Cursedness"
that can activate during the game.

I still want a way to control the colour and metallic etc, so it blends depending on moving objects or controller input, almost like screen aberration or something.

Speaking of screen: it could be good to also look at the screen space itself for glitchy artefacts, because that is a different level of the pipeline.