Scriptable Objects
Setup all the dialogue scriptable objects with the old dialogue
There is currently an issue with the bits that handle the player response.
There is a coroutine at the end which will handle special logic for the player response, and it is currently not working with how the update is updating/resetting the bools for the button clicks.