Dialogue Event Handling
Set up a few things to handle the events triggered by dialogue.
It works with a few things working together:
I have enums to name specific events that can be assigned in dialogue data SO's
They are set up as arrays, so that more than one event can be assigned under a dialogue unit.
Then I have a DialogueEventManager that takes those given event names (enum arrays)
and will then run the methods that correspond to them.
The DialogueEventManager is also linked to the sceneswitcher because that is some of the basic functions that happens at the end of a dialogue.
So, as the need for more specific events arise, I can add the enum to the list for the name of the event
Add in a case and a method in the DialogueEventManager.