Scriptable Objects for Char Stats
Still working on making the dialogue system work.
I have switched out the Character Stats class for a scriptable object, and trying to work out how that stats will be modified during the game.
SOs means the stats wont reset every scene, which would be no good.

Im super stuck at the moment scratching my head, because the queries in the meeting scene is not handling the GameConditions checks as expected. The conditions are clearly defined in the scriptable object for the queries, as well as in the dialogue units SO. But for some reason it returns false when it checks for the gamecondition. The only gamecondition in the query is the dialogue line, and for some reason the check thinks it is false. I have checked everywhere to ensure that the "DialogueLine" gamecondition is indeed the expected value, and it is. But still, the !IsGameMatch check returns false. I DONT UNDERSTAND