Scriptable Object Shenanigans
I may have broken the game, I replaced the CharacterStats class with scriptable objects, so I could persist the data.
I discovered you cannot make dictionary data persist through use of SO's
SO i used Ints to store the Values, and then made some methods in the StatsManager to convert dictionary data to correspond to the int values, and vice versa. When the script unloads, its stores the ints, and when a new scene loads, i Updates the dictionaries again to the correct values.

WHAT IM WORRIED ABOUT: I messed with a value in the old CharacterDataSO "infectionLevel", which was a float. I made it an int so it can work bteter with my current stats. But it may have broken some scripts and methods that use that variable. SO will nee to watch out for broken things down the line...