SOs persisting data
holy craparoni i fixed it
Seriously though, what a pain in the ass.
It turns out the SOs were restting to default values because they were not present in the next scene. Weird
I just actually added in the stats manager in every scene with the list of SOs in it,
That seemed to fix it.