Speaker Criteria
So moving on to... branching dialogue
If characters lie, I want other characters to respond differently.

Need to actually make a bool check for SpeakerCriteria in the query. All it does at the moment, is find a dialogue unit that matches the query criteria... BUT it doesn't actually know if such a speaker even exists in the first place. SO it needs a check before even running the query, same as it does for IsGameMatch. Something like IsSpeakerMatch. So if it returns false, we just ksip to the next query and return.