Finally....
So I figured out the issue.
I was checking for matching speakers, to match with all given criteria. So I wanted to only continue if EVERY criteria was true. What was happening, is it found a speaker for every criteria, no problem. Except it wasn't the SAME speaker lol. So it was almost always true.
Had to add logic to check that every matched criteria also has the SAME speaker across ALL matches.