Up to meeting scene 2...
Fixed a bunch of problems with data not flowing well between scenes, not initialising properly, etc.
So far the dialogue system seems to mostly work as desired.
I can create different conditions with different responses for characters depending on the conditions.
I have some thoughts for things that could have been better... For one thing, the player response doesn't have the same flexibility as the dialogue units. As in, I can't set up special conditions for the player response, for example, if the team is cursed to some degree, the player might have specific responses, or if the team is not cursed at all, theres another option. Because if there were two cases like this, I could actually decide the player has lied or not...