Dialogue Data Setups/Resets etc
Think I fixed the issues with stats not being reset and initialised properly for scene changes.
Had to just carefully make sure it gets set and reset at the appropriate times in the game.
The tricky thing is that data gets passed from the character manager to the stats manager, and the stats manager sets up dictioaries, manages the updates/modifications to the stats values. So its important not to lose track of where the data gets copied from, when it should set those values or when it should reset to defaults, etc.
I also had some issues with how events gets called from dialogue events, specifically switching to next dialogue sequence...