Player response Options
Made changes to the scripts so that there can be queries to select different player response options.
So for example, if there is Curse on the team, it will queue different options for player response
than it would if there was no curse.
It also means that in the different responses, I can actually know what game condition queued the responses, so that gives me context to know if the player is lying or not.
For example, if there is curse, and the player says the curse is gone, I can trigger an event that the player has lied.