Elimination Vote Scene
working on a scene for elimination voting.
Its all set up, but there might be issues with indexes getting mixed up, perhaps a difference in the order of characters in different scenes?
Like, indexes of characters and their stats, (VoteTarget etc) might not correspond correctly in the Vote Scene.
Check this...