GudVsCursed
Included some stats that can start tracking if the team is aligned more with Gud or Cursed.
Can align characters with Gud side or Cursed Side by selecting "SpeakerSides" in dialogue events
Then there is also a game check for GudBeatsCursed, which can be -1 (default, meaning theyre the same)
or 0, which means cursed wins, or 1, which means Gud wins.
So this could be useful for choosing dialogue options for characters.
Characters with influence aligned with gud can maybe affect others in different ways
than characters with influence aligned with cursed.
I have a feeling i should look at how vote targets are calculated again, to factor in the group alignment or individual character's alignment