Vox Emotes (and UI tweak)
Selecting Vox emote options in dialogue units.
I would very much like to redo all the sounds for the dialogue vocal sounds, but that must come later
I also replaced a button in the debug game: so that you press x to return in BOTH cases. It was A to return and then X to return the second time... which is a bit confusing. So we use A for entering, x to return, as a convention.