New Proxy Rig and Recording
Been working my buttus off on getting a new proxy rig to work
Everything works ok, would still need to set up
I tried a physics based movement but it got complicated (these were changes on the previous branch, still deciding if i want to use physics or not...)
Now did transforms based movement.
There is currently an issue with the recording vs playback of the root positions. I need a better way to control if it records/plays, as in, overrides recorded data or uses the recorded data. I need solid logic or conditions to control that