Looks and VOX clips
The looks and vox clips were quite mixed up.
I just fixed some index problems in the camera, looks, and audio scripts, which handles the logic for playing the right voices, setting the looks, and switching cameras, according to indexes of characters.
The scripts had npc2, 3, 4 for chars at index 1,2,3, which could already create issues. The scripts also had extra npc indexes, like 4 and 5, which would be redundant since there are only 3 npcs if there is a player, demon, and cursed char.
The dialogue SO's might need to be checked to make sure all the indices are set correctly for each dialogue unit (a.i. the looks positions, camera positions)