STRUGGGGLING: Feet Planting and Root Movement
Been working on little script tweaks to keep one foot planted if other foot is lifted.
The way it is scripted is not sophisticated enough and glitchy(not in a good way)
It sorta works though, could do with more tweaking.
I wanted to make it so feet move in inverse to the root movement, so that it looks like it stays in place if the root object moves. I must be missing something, because this is not working at all.
I am out of touch with how the movement works, so I might be missing something in how the updates depend on recorded positions and use temporary vectors to store positions etc etc. WIll have to look super hard at all these intertwined processes...