FootPlanting+RootMovement
The foot-planting in conjunction with the root movement sort-of works.
The inverse offset works, it just behaves weird with the way the rig works. The feet end up lifting into the air, because the target object is so far that the foot is basically lifting into the air.

I am wondering if it is actually better to NOT plant the feet. The unbelievable physics of hovering in the air, sliding along the floor, etc, can be jarring but potentially be something to play with?