Spinning Challenges
Still trying to make the spinning work
the tricky bit is that the rig of the characters are independant objects,
but they are copying armature objects that are in a hierarchy, so that moves objects via parent-child relationships.
copyiing objects that are rotated as children around paren't rotation, but then recording "localPostitions" means the rigs don't know how to move their objects as if they are rotating around a parent.
Ok, so my thought here is to use world space, but that might mess everything up...