Player Cursed
The player now gets an average of the other characters' cursedness.
I think it adds something ocol when the cursed behaviour is interactive and available for the player to manipulate.
This will be where the heart of this games "experiment" can really get pushed.
How does the cursedness affect player experience? Is it something interesting to play with? Can I push the meaning?