jump glitch fixed
there was a jump glitch which was because when we switch between dance moves, it adds the current position of the root in world space as an offset to every other character's root.
I changed it so it stores the "initial" position of the root (stored at the start), adding that as an offset instead of some updated position which might be mid-jump or moved into a different x,z position, which is NOT what we want.