Timing, Beats, Sync
Lots of changes with the Clock Timer

It samples the actual audio, and then calculates intervals accordingly, so it keeps an accurate "beat" interval that can now be used by all the scripts that depends on that.

I actually just changed the way the variables THAT ARE ALREADY GETTING USED gets calculated, so I dont actually have to go trhough all the other scripts and change everything, right...

Might still have to check for issues.