Lots of SHaderBender Experiments
TWeaking and tinkering... looking for ways to make the geometry displacement and irridescent colours be driven by different things in the game.

Obvious choice was using the beat, or the frequency bands.
Took a bit of inagling to access the frequency data (spectrum data).

Theres something interesting about how maybe different parts of my body responds to different frequencies of music.
AT the moment, I mapped frequencies like the pelvis to bass, and higher frequencies are like hands and the head.

Could play with this some more.

I have noticed that the head isn't affected very good, might be because the origin of the bone is kinda in the neck, so the falloff happens from there. Could replace it with something more in the middle of the head to control those shader parameters for head geaometry displacement.