New GAmeplay Loop
Changed game to focus on two scenes of making dances + copied dancing...
SHould rethink how the curse srpeads, as there are 4 rounds and 6 chars,
so how does the curse spread if not 1 by 1?
It could actually be good to not rely on a 1 by 1 progression!
The next consideration, is what it will look like and how it will work to show characters reacting to the curse.
Will it be an inserted frame within the scene, as conventionally in games for char dialogue?
Or could it be done with cinematography, cameras on characters WHILE they are dancing?
What will happen with the cutscene drama music, will this lose some of that cool drama?
Think about how important the storytelling layer is for supporting the "queer" dancing. How does it work? does it reinforce, support, or is it superfluous