Fixed the NaN issues
there was an issue with calculations for shaderbender (and the debug 5-bar visualisation)
because some calculations returned NaN
it turns out the issue was with the infection calculations: at some point the "average infection" gets calculated by dividing the total infection by the amount of infected characters. but in the first scene it didn;t add the infected character to the "infected characters" list, so it tried to divide by zero to get the "average infection".
The player characer uses that average infection to control the shading variables, but because it was NaN, a lot of issues ensued.
Now that the cursedness works, I want to adjust how the infection spreads.
There are four rounds, but 5 characters to get cursed. So the way they "get cursed" should be spread over those four rounds. I think it should rely on if/ how much the player "debugs" the cursedness. I think I need a spreadsheet for this or something...