Aligner Game - Controller input Changes
Just reworked the way the controller input measurements work
I am using dot product and direction to get a discrepency of direction values, instead of a distance measured between two vectors. This means even if the controller is minimally pushed, all that matters is the direction of that vector.

On top of that, I use the magnitude of the thumbstick input, so how much it is pushed to the edge, as an extra factor. If it is not pushed at all, then the alignment is at 0 (most discrepency). Pushing it all the way to the edge gives it a value of 1, so the "alignment" amount is multiplied by 1. So that direction value gets multiplied by this factor.

I would also like to use these magnitude values as extra inputs for the sound. For example, it could control the low cut filter, or a volume LFO (I have 2 thumbstick magnitudes to use)