Normaline Audio Stuff
Just messing around with some more audio effects responding to manipulation of the thumbsticks
I want to get it to a point where it produces really fun and interesting results when players explore the circle area.
At the moment, there are some funky effects, but it is not producing a very interesting range of results.
I can see this working better in the unreal metasound system, and setting up a custom soundmachine designed specifically for this game.