Freq Dancing
Still working on the frequaliser dance script
Added some better logic for the lerping. using a max and min frequency for each band, which gets updated to set a max and min target, then inverse lerp between those, instead of just between 0-1.
There is still some work that can be done to set targets better, for example, the arms are mostly stretched outwards and it doesn't animate very well.