MAking A Rave Demon
I have been working on the render texture shader.
THe plan is that the literal surface of the game comes alive, so what was the 2d screen might become 3d entity, or perspective gets messed up, or something...
Im just tryng things out
Right now, I have created a literal demon face that the screen morphs into, then the texture looks like its projected over the surface.
Its ok... but i dont love it
Ive been thinking of "boots the house down" or like, tear the house up, or whatever, anyway, I like the idea of the house literally coming down, the game falling apart, or something like that.
Maybe we dont literally need to see a "demon"
I think I can think deeper about what the "demon" is?
In a way, it represents the fear of uncomprehesable/alien entities in the machine, what is out of our control
And at the same time, it represents possibilities, the liberatory power of disruption, of dancing with the machine, letting go of control.