TRiangulatedScreen
HAvent achieved much, except to replace the "grid-like" screen with a bit of a randomish triangulated mesh.
I have three different meshes, I could swap between them, so the triangles look a bit more animated...

Been trying to understand how the geometry shader is working.
It is shrinking the triangles, but it eems to add triangles on top of each other, like it retains triangles and keeps adding on top of each other, for some reason